# services/core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.EmployeeProfileResponse:
    properties:
      plans:
        items:
          $ref: '#/definitions/dto.PlanResponse'
        type: array
      skills:
        items:
          $ref: '#/definitions/dto.SkillResponse'
        type: array
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.PlanResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      creation_type:
        type: string
      description:
        type: string
      employee_id:
        type: string
      generation_status:
        type: string
      id:
        type: string
      progress:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.ProfileResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      position:
        type: string
      role:
        type: string
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  dto.RegisterResponse:
    properties:
      id:
        type: string
    type: object
  dto.SkillResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.UpdatePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      manager_id:
        type: string
      position:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Mandzhonson
  description: REST API для SkillTracker.
  title: SkillTracker API
  version: "1.0"
paths:
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Регистрация нового пользователя
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - Auth
...
  /manager/employees:
    get:
      consumes:
      - application/json
      description: Возвращает список сотрудников, закрепленных за менеджером
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список сотрудников менеджера
      tags:
      - Manager
  /manager/employees/{employee_id}:
    get:
      consumes:
      - application/json
      description: Возвращает расширенный профиль сотрудника с навыками и планами
      parameters:
      - description: ID сотрудника
        in: path
        name: employee_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EmployeeProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить профиль сотрудника (для менеджера)
      tags:
      - Manager
  /manager/employees/{employee_id}/avatar:
    get:
      consumes:
      - application/json
      description: Возвращает аватар сотрудника
      parameters:
      - description: ID сотрудника
        in: path
        name: employee_id
        required: true
        type: string
      produces:
      - image/jpeg
      - ' image/png'
      - ' image/gif'
      responses:
        "200":
          description: Аватар сотрудника
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить аватар сотрудника (для менеджера)
      tags:
      - Manager
...
  /users/me:
    get:
      consumes:
      - application/json
      description: Возвращает профиль авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить профиль текущего пользователя
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Обновляет данные профиля пользователя
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить профиль
      tags:
      - User
  /users/me/avatar:
    delete:
      consumes:
      - application/json
      description: Удаляет аватар пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить аватар
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Возвращает аватар пользователя
      produces:
      - image/jpeg
      - ' image/png'
      - ' image/gif'
      responses:
        "200":
          description: Аватар пользователя
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить аватар
      tags:
      - User
    put:
      consumes:
      - multipart/form-data
      description: Загружает аватар пользователя
      parameters:
      - description: Файл аватара
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузить аватар
      tags:
      - User
  /users/me/password:
    patch:
      consumes:
      - application/json
      description: Изменяет пароль пользователя
      parameters:
      - description: Пароли
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Сменить пароль
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"